+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+### Added
+- `add_transaction` method
+
 ### Changed
 - Removed `offset` param for paginated calls, use `limitstart` instead.